# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PMS Hotel Management System API
  description: |
    Complete REST API documentation for the Property Management System.
    
    ## Authentication
    All endpoints (except login) require a Bearer token in the Authorization header.
    
    ## Base URL
    - Development: http://localhost:3001/api
    - Production: https://api.yourhotel.com/api
    
    ## Error Codes
    - 400 - Bad Request
    - 401 - Unauthorized
    - 403 - Forbidden
    - 404 - Not Found
    - 500 - Internal Server Error
  version: 1.0.0
  contact:
    name: API Support
    email: support@grandview.com
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.grandview.com/api
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /rooms:
    get:
      tags:
        - Rooms
      summary: Get all rooms
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: floorId
          in: query
          schema:
            type: integer
        - name: roomTypeId
          in: query
          schema:
            type: integer
        - name: branchId
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of rooms
    post:
      tags:
        - Rooms
      summary: Create new room
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - branchId
                - roomNumber
                - roomTypeId
                - floorId
              properties:
                branchId:
                  type: integer
                roomNumber:
                  type: string
                roomTypeId:
                  type: integer
                floorId:
                  type: integer
                status:
                  type: string
                notes:
                  type: string
      responses:
        '201':
          description: Room created
  /rooms/{id}:
    get:
      tags:
        - Rooms
      summary: Get room by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Room details
    put:
      tags:
        - Rooms
      summary: Update room
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                roomNumber:
                  type: string
                roomTypeId:
                  type: integer
                floorId:
                  type: integer
                status:
                  type: string
      responses:
        '200':
          description: Room updated
    delete:
      tags:
        - Rooms
      summary: Delete room
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Room deleted
  /rooms/availability/check:
    get:
      tags:
        - Rooms
      summary: Check room availability
      security:
        - bearerAuth: []
      parameters:
        - name: checkIn
          in: query
          required: true
          schema:
            type: string
        - name: checkOut
          in: query
          required: true
          schema:
            type: string
        - name: roomTypeId
          in: query
          schema:
            type: integer
        - name: guests
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Available rooms
  /rooms/types/list:
    get:
      tags:
        - Rooms
      summary: Get room types
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of room types
  /rooms/types:
    post:
      tags:
        - Rooms
      summary: Create room type
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - typeName
                - basePrice
              properties:
                typeName:
                  type: string
                description:
                  type: string
                capacity:
                  type: integer
                basePrice:
                  type: number
                maxOccupancy:
                  type: integer
                bedType:
                  type: string
                sizeSqFt:
                  type: integer
      responses:
        '201':
          description: Room type created
  /rooms/types/{id}:
    get:
      tags:
        - Rooms
      summary: Get room type by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Room type details
    put:
      tags:
        - Rooms
      summary: Update room type
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                typeName:
                  type: string
                description:
                  type: string
                basePrice:
                  type: number
      responses:
        '200':
          description: Room type updated
    delete:
      tags:
        - Rooms
      summary: Delete room type
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Room type deleted
  /rooms/amenities:
    get:
      tags:
        - Rooms
      summary: Get room amenities
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of amenities
    post:
      tags:
        - Rooms
      summary: Create amenity
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - amenityName
              properties:
                amenityName:
                  type: string
                amenityType:
                  type: string
                description:
                  type: string
                icon:
                  type: string
      responses:
        '201':
          description: Amenity created
...
  /rooms/status-board:
    get:
      tags:
        - Rooms
      summary: Get room status board
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Status board data

  # ============================================================
  # GUESTS
  # ============================================================
...
  /roles:
    get:
      tags:
        - Roles
      summary: Get all roles
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of roles
    post:
      tags:
        - Roles
      summary: Create role
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - roleName
              properties:
                roleName:
                  type: string
                roleDescription:
                  type: string
      responses:
        '201':
          description: Role created
  /roles/{id}:
    get:
      tags:
        - Roles
      summary: Get role by ID with permissions
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Role details
    put:
      tags:
        - Roles
      summary: Update role
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                roleName:
                  type: string
                roleDescription:
                  type: string
                isActive:
                  type: boolean
      responses:
        '200':
          description: Role updated
    delete:
      tags:
        - Roles
      summary: Delete role
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Role deleted
  /roles/{id}/permissions:
    post:
      tags:
        - Roles
      summary: Update role permissions
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - permissions
              properties:
                permissions:
                  type: array
                  items:
                    type: integer
      responses:
        '200':
          description: Permissions updated

  # ============================================================
  # PERMISSIONS
  # ============================================================
...